Add files via upload
This dataset has COVID cases till Sep 13 which was initially supposed to be the end date of stage 4 restrictions. With this dataset, I have predicted what would have been the daily number of COVID cases from the start of stage 4 restrictions i.e 2nd August till the 13th of September.